Refactor: Extract components according to review comments
