+++ src/App.jsx
@@ -1,33 +1,19 @@
 import { useState } from 'react';
 import { restaurants } from './mock';
 import Layout from './components/layout/Layout';
-import Restaurant from './components/restaurant/Restaurant';
-import RestaurantTabs from './components/restaurant/RestaurantTabs';
+import RestaurantContainer from './components/restaurant/RestaurantContainer';
 import { ThemeProvider } from './contexts/ThemeContext';
 import { UserProvider } from './contexts/UserContext';
 import { CartProvider } from './contexts/CartContext';
 import './reset.css';
 
 function App() {
-  const [activeRestaurantId, setActiveRestaurantId] = useState(restaurants[0]?.id);
-  const activeRestaurant = restaurants.find(r => r.id === activeRestaurantId);
-
   return (
     <ThemeProvider>
       <UserProvider>
         <CartProvider>
           <Layout>
-            <RestaurantTabs
-              restaurants={restaurants}
-              activeId={activeRestaurantId}
-              onTabClick={setActiveRestaurantId}
-            />
-            {activeRestaurant && (
-              <Restaurant 
-                restaurant={activeRestaurant} 
-                multiplier={3} 
-              />
-            )}
+            <RestaurantContainer restaurants={restaurants} />
           </Layout>
         </CartProvider>
       </UserProvider>
