+++ src/components/user/UserAuth.module.css
@@ -1,0 +1,86 @@
+.userAuth {
+  display: flex;
+  align-items: center;
+}
+
+.loginButton, 
+.logoutButton {
+  padding: 0.5rem 1rem;
+  border-radius: 4px;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.loginButton {
+  background: var(--button-primary-bg);
+  color: white;
+  border: none;
+}
+
+.loginButton:hover {
+  background: var(--button-primary-hover-bg);
+}
+
+.logoutButton {
+  background: var(--button-danger-bg);
+  color: white;
+  border: none;
+}
+
+.logoutButton:hover {
+  background: var(--button-danger-hover-bg);
+}
+
+.userInfo {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.username {
+  font-weight: 600;
+  color: var(--text-primary);
+}
+
+.loginForm {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.loginInput {
+  padding: 0.5rem;
+  border: 1px solid var(--border-color);
+  border-radius: 4px;
+  font-size: 0.9rem;
+  background-color: var(--bg-primary);
+  color: var(--text-primary);
+}
+
+.cancelButton {
+  padding: 0.5rem 1rem;
+  border-radius: 4px;
+  background: var(--button-secondary-bg);
+  color: white;
+  border: none;
+  font-weight: 600;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+.cancelButton:hover {
+  background: var(--button-secondary-hover-bg);
+}
+
+/* Адаптивные стили для мобильных устройств */
+@media (max-width: 768px) {
+  .loginForm {
+    flex-wrap: wrap;
+  }
+  
+  .userInfo {
+    flex-direction: column;
+    align-items: flex-end;
+    gap: 0.25rem;
+  }
+}
